v4.0.0 b1
more work

+++ levelledmobs-plugin/src/main/kotlin/io/github/arcaneplugins/levelledmobs/customdrops/CustomDropsHandler.kt
@@ -1279,7 +1279,7 @@ class CustomDropsHandler {
                 val result: PlayerLevelSourceResult = main.levelManager.getPlayerLevelSourceNumber(
                     info.mobKiller, info.lmEntity!!, variableToUse
                 )
-                levelToUse = if (result.isNumericResult) result.numericResult else 1
+                levelToUse = if (result.isNumericResult) result.numericResult.toInt() else 1
                 info.playerLevelVariableCache[variableToUse] = levelToUse
             }
